idstrap_detector/requirements.txt
streamlit>=1.25.0
opencv-python
ultralytics>=8.0.0
roboflow
torch>=1.13.0
numpy
pillow
